refactor(party): partyroom.crew_count 캐시 카운터 완전 제거 (#360)
이벤트 구동 카운터는 이벤트 없는 상태 변경(V38 collapse·운영 수동 SQL)에서 드리프트하며
#358 어드민 크루 수 불일치의 근원이었다. #359 로 모든 소비자가 라이브 COUNT 전환을 마쳐
(고객 로비·신/구 어드민·입장검증 전수 재검증: reader 0) 쓰기 전용 유지비용과 "그럴듯한
이름의 죽은 데이터" 함정만 남아 제거한다.

- V39: alter table partyroom drop column crew_count
- PartyroomData.activeCrewCount 필드 제거
- PartyroomRepository: increment/decrement/resetCrewCount 삭제 (touchLastActivity 존속)
- PartyroomCounterListener → PartyroomActivityListener 개명·축소: ENTER/EXIT/재생 이벤트의
  lastActivityAt touch 만 담당 (terminate/close reset 핸들러는 대상 소멸로 삭제)
- 테스트 재편: ActivityListenerIT(touch 단언), 카운터 동시성 IT 삭제, RaceIT "100 동시 enter"
  단언을 카운터→라이브 활성 crew=1 로 격상, 의존 5파일 정리

미래 대규모 스케일에서 캐시 카운터가 필요해지면 reconcile 동반 설계로 재도입.
