+++ app/src/test/java/com/pfplaybackend/api/party/domain/entity/data/PartyroomDataTest.java
@@ -41,7 +41,6 @@ class PartyroomDataTest {
             PartyroomData p = newPartyroom();
             assertThat(p.getStatus()).isEqualTo(PartyroomStatus.ACTIVE);
             assertThat(p.getDisplayFlag()).isEqualTo(DisplayFlag.NORMAL);
-            assertThat(p.getActiveCrewCount()).isZero();
             assertThat(p.getLastActivityAt()).isNull();
             assertThat(p.isActive()).isTrue();
             assertThat(p.isSuspended()).isFalse();
